pyproject.toml
[project]
name = "tender-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "Murilo Seidenstucker",email = "murilocampos.004@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "tzdata (>=2025.3,<2026.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "testcontainers (>=4.14.1,<5.0.0)"
]

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
lint = "black --check --diff . && isort --check --diff ."
pre_test = "task lint"
test = "pytest -s -x --cov=src -vv"
post_test = "coverage html"
docker_build = "docker-compose up -d --build"
docker_up = "docker-compose up -d"
docker_down = "docker-compose down"
